requirements.txt
streamlit
streamlit-extras
azure-cognitiveservices-speech
azure-ai-translation-document
azure-ai-translation-text
openai
python-dotenv